webaudio/the-audio-api/the-audioparam-interface/method-linearRampToValueAtTime.html

<!doctype html>
<html>
<head>
	<title>AudioParam interface - LinearRampToValueAtTime tests | WebAudio</title>

	<script type="text/javascript" src="../../../resources/testharness.js"></script>
	<script type="text/javascript" src="../../../resources/testharnessreport.js"></script>
	
	<script type="text/javascript" src="../../js/WebAudioTestHarness.js"></script>
	<script type="text/javascript" src="../../js/WebAudioTestHarness-custom.js"></script>
	<script type="text/javascript" src="audioparam-specific.js"></script>
</head>
<body>
	<div id="log">
	</div>

	<script type="text/javascript">
		/*
		LinearRampToValueAtTime without previous scheduled-event.
		Specification says that LinearRampToValueAtTime ramps from the previous scheduled-event.
		If there is no previous scheduled-event then the AudioParam does not know when the slope should
		start.  So in this case, the method acts exactly like setValueAtTime - and the value instantaneously
		changes from the previous value to the new value.
		*/
		(function() {
			executeTest("LinearRampToValueAtTime without previous scheduled-event", arrangeFn, assertFn);

			function arrangeFn(ac, audioParam) {
				audioParam.value=0;

				//Schedule linearRampToValueAtTime to increase value to 1 after 250 samples.
				audioParam.linearRampToValueAtTime(1, this.convertSamplesToSeconds(250));
			}

			//Assert that the value changes at the correct time.
			function assertFn(actualData) {
				assert_approx_equals(actualData[0], 0, 0, "Expect start of output to be 0");
				assert_approx_equals(actualData[249], 0, 0, "Expect output to be 0 before change has finished");
				assert_approx_equals(actualData[250], 1, 0, "Expect output to be 1 after change has finished");
			}
		})();

		/*
		LinearRampToValueAtTime with previous scheduled-event.
		*/
		(function() {
			executeTest("LinearRampToValueAtTime with previous scheduled-event", arrangeFn, assertFn);

			function arrangeFn(ac, audioParam) {
				//The previous scheduled event.
				audioParam.setValueAtTime(0, 0);

				//Schedule linearRampToValueAtTime to ramp from 0 to 1 over 250 samples.
				audioParam.linearRampToValueAtTime(1, this.convertSamplesToSeconds(250));
			}

			//Assert that the value changes at the correct time.
			function assertFn(actualData) {
				assert_approx_equals(actualData[0], 0, 0, "Expect start of output to be 0");
				assert_approx_equals(actualData[125], 0.5, 0, "Expect output to be 0.5 half-way up");
				assert_approx_equals(actualData[250], 1, 0, "Expect output to be 1 after ramp has finished");
			}
		})();

		/*
		Multiple linearRampToValueAtTime calls.
		*/
		(function() {
			executeTest("Multiple linearRampToValueAtTime calls", arrangeFn, assertFn);

			function arrangeFn(ac, audioParam) {
				//The previous scheduled event.
				audioParam.setValueAtTime(0, 0);

				//Schedule linearRampToValueAtTime to increase value to 1 over 250 samples.
				audioParam.linearRampToValueAtTime(1, this.convertSamplesToSeconds(250));

				//Schedule back to 0 over the next 500 samples.
				audioParam.linearRampToValueAtTime(0, this.convertSamplesToSeconds(250+500));
			}

			//Assert that the value changes at the correct time.
			function assertFn(actualData) {
				assert_approx_equals(actualData[0], 0, 0, "Expect start of output to be 0");
				assert_approx_equals(actualData[125], 0.5, 0.01, "Expect output to be 0.5 half-way up");
				assert_approx_equals(actualData[250], 1, 0.01, "Expect output to be 1 before down-ramp starts");
				assert_approx_equals(actualData[500], 0.5, 0.01, "Expect output to be 0.5 half-way down");
				assert_approx_equals(actualData[750], 0, 0.01, "Expect output to be 0 after down-ramp is complete");
			}
		})();
	</script>
</body>
</html>
